+++ Question5/src/main/java/com/revature/question5/Map1.java
@@ -2,13 +2,14 @@ package com.revature.question5;
 
 import java.io.IOException;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 import org.apache.hadoop.mapreduce.Mapper.Context;
 
-public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
+public class Map1 extends Mapper<LongWritable, Text, Text, DoubleWritable>{
 	@Override
 	public void map(LongWritable key, Text text, Context context) throws IOException, InterruptedException{
 		String line = text.toString();
@@ -22,10 +23,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Primary School: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Primary School: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Primary School: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Primary School: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -39,10 +40,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Secondary School: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Secondary School: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Secondary School: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("Government Expediture (% of GDP per Capita) per Student: Secondary School: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
@@ -56,10 +57,10 @@ public class Map1 extends Mapper<LongWritable, Text, Text, IntWritable>{
 			for(int x = 4; x < rows.length; x++) {
 				try {
 					if(rows[x].length() == 0 && year >= 2000) {
-						context.write(new Text("Government Expediture (% of GDP per Capita) on Education in General: " + CountryName + ", " + year), new IntWritable(-1));
+						context.write(new Text("Government Expediture (% of GDP per Capita) on Education in General: " + CountryName + ", " + year), new DoubleWritable(-1));
 					} else if (rows[x].length() != 0 && year >= 2000){
-						Integer Data = Integer.parseInt(rows[x]);
-						context.write(new Text("Government Expediture (% of GDP per Capita) on Education in General: " + CountryName + ", " + year), new IntWritable(Data));
+						Double Data = Double.parseDouble(rows[x]);
+						context.write(new Text("Government Expediture (% of GDP per Capita) on Education in General: " + CountryName + ", " + year), new DoubleWritable(Data));
 					}
 					year++;
 				} catch (ArrayIndexOutOfBoundsException e) {
